requirements.txt
dotenv
fastapi
langchain
langchain-community
langchain-core
langchain-experimental
langchain-aws
langchain-postgres
langchain-mistralai
langgraph
langgraph-checkpoint-postgres
lifelines
matplotlib
pandas
psycopg[binary,pool]
pydantic
python-dotenv
pwdlib
pyjwt[crypto]
requests
scipy
seaborn
statsmodels
sqlalchemy
uvicorn